// frontend-foodly/src/types/auth.types.ts
import type { UserType } from './user.types';

export interface AcccessTokenData {
    user_id: number;
    user_fullname: string;
    user_email: string;
    user_type: UserType;
    iat: number;
    exp: number;
}

export interface RefreshTokenData {
    user_id: number;
    user_email: string;
    iat: number;
    exp: number;
}